Changed evalcircles.m to accept parameter-grid. Had to write a slightly modified version of getcircles: I called it findcircles. We should keep only one of those. So instead of calling getcircles, evalcircles it now calles findcircles.
Also, extracted the code the removes circles within circles to new function.

Added a new external lib, that contains the classical Haskell functions. Comes really handy when you want to aggregate results.

Added a script that shows how to use alexnet.m. Due to local path references, it won't be executable on your machine. So it's just serves the purpose of a code snippet to look up when needed.